flask-template/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pre-request==2.1.5
pycparser==2.21
PyMySQL==1.1.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1

PyJWT~=2.8.0